fixed figure label as well
