+++ elements/stop-note/lib/stop-note-proposed.js
